C_Nebula/final-pjt-back/requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
cors==1.0.1
cryptography==38.0.3
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==3.2.13
django-allauth==0.51.0
django-cors-headers==3.13.0
django-extensions==3.2.0
djangorestframework==3.14.0
executing==1.0.0
filelock==3.8.0
future==0.18.2
fuzzywuzzy==0.18.0
gevent==22.10.2
greenlet==2.0.1
idna==3.4
ipython==8.4.0
jedi==0.18.1
Levenshtein==0.20.8
matplotlib-inline==0.1.6
oauthlib==3.2.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
PySocks==1.7.1
python-Levenshtein==0.20.8
python3-openid==3.2.0
pytz==2022.2.1
rapidfuzz==2.13.2
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.5.0
tldextract==3.4.0
traitlets==5.3.0
urllib3==1.26.12
wcwidth==0.2.5
zope.event==4.5.0
zope.interface==5.5.1
